Update README.md
My name is Nadezhda. I want to work as a Prompt engineer.
I have completed advanced training in PROMT-ENGINEER (the art of communication with a neural network) and am a specialist. I know and understand how to write a request in such a way as to get the desired result. I objectively evaluate the results, double-check the facts, and know how to find their confirmation or refutation.
I develop a unique prompt and am also attentive to detail.
I quickly find information in Russian- and English-language sources and immerse myself in different areas.
Ready to complete tasks on a personal computer.
I expect to begin my path as a Prompt engineer in a team, under the guidance of professionals..